==> code/2_1_EnvTrends_caplter-birdcommunitydynamics.Rmd <==
---
title: "CAP Bird Community Dynamics - Testing and Visualizing Environmental Patterns and Trends"
author: "Jeffrey Haight et al."
date: "`r Sys.Date()`"
output:
  word_document: default
  html_document: default
---

Code for testing and visualizing spatial-temporal trends in, and relationships among, environmental conditions

```{r setup, include=FALSE}
knitr::opts_chunk$set(
  echo = FALSE,
  warning = FALSE, 
  include = TRUE
  )
rm(list = ls())
gc()
```

```{r packages, message = FALSE, echo = TRUE}
library(beepr)
library(vegan)

library(sf)
library(terra)
library(tidyverse)
library(tidyterra)  # tidyverse methods for SpatRaster and SpatVector objects (including ggplot functions)
# library(tidycensus)
#library(tmap)
#library(stars)
library(ggcorrplot)
library(GGally)
library(gghighlight)
library(RColorBrewer)
library(lterpalettefinder)
library(scales)

# Census data settings
options(tigris_use_cache = TRUE)
#census_api_key("fd53ba9b6b6ab4fa5a0fcf329401ae7d3184dc6a", install = TRUE) # set the Census API key


# for statistical tests and modeling
library(ggpubr)     # for adding correlations and p-values to ggplot
library(lme4)
library(glmmTMB)
library(lmerTest)   # for getting p-values out of the lmer models
library(MuMIn)
library(performance)
#library(regclass)
library(GLMMadaptive)   # for calculating VIF for 'glmmTMB' models
library(ggeffects)    # for plotting glmm effects in ggplot
```
```{r color palettes} 

  # import some color palettes to use for visualizing seasons
  load("./data/arizonapalettes.RData")

palette_ggdemo(asu)
palette_ggdemo(wildflower)
palette_ggdemo(biocrust)
palette_ggdemo(sonoran)

# Palettes for the four seasons based on local colors
(pal_season <- c(
  "#2f3ca3", # arroyo lupine
  # "#4ab7c4",
  # "#7291fb",
  # "#70a112", # moss green
  # "#c8ff08", # lichen green (too bright against a white background)
  "#78be20", 
  "#d10073",  # prickly pear pink #1
  # "#ff8701", # poppy orange
  "#ffc901", # palo verde yellow
  "#000000"
)) #%>% palette_ggdemo()

(pal_season_asu <- c(
  "#00a3e0", 
  "#78be20", 
  "#e74973", 
  "#ff7f32",
  "#000000"
)) #%>% palette_ggdemo()

# (pal_season_light <- c(
#                   "#7291fb",
#                   "#f2ff0d",  # brittlebush yellow
#                   "#ff23c1"   #prickly pear pink #2
#                   )) %>% palette_ggdemo()


```


# Import Data
```{r, include = FALSE}
list.files("./data")

# This contains data assembled by the '2_1_StatisticalModelingSetup_CAPbirds' file
load("./data/modelinputs_CAPbirds2024.RData")

# comparing the bird survey temporal extent to a broader trend
# filter all four season of data to just the sites and years surveyed
    env.model <- data.env %>% 
      # filter(season == "3_summer") %>% 
      filter(site_code %in% unique(data.env.model$site_code))%>% 
      filter(survey_year %in% unique(data.env.model$survey_year)) %>%
      mutate(year_std = (survey_year - mean(survey_year))/sd(survey_year))
    
    env.model$loc_type <- as.factor(env.model$loc_type)
```


# COLLINEARITY BETWEEN ENVIRONMENTAL VARIABLES
### Check for bivariate collinearity among predictor and response variables
```{r across both seasons}
var.cont <- env.model %>% filter(!season %in% c("annual")) %>% 
    select(c(NDBI, NDBI_prevsum, #NISI, NDISI, ENDISI, 
             NDVI, NDVI_prevsum, #SAVI, MNDWI, 
             LST, LST_prevsum, 
             ppt_sum, ppt_sum_prevsum,
             #temp_max, temp_max_prevsum, 
             temp_min, temp_min_prevsum,
             # abundance, rich, 
             survey_year
             ))
corplot.all <- ggcorrplot(cor((var.cont), use = "complete.obs", method = "spearman"), 
           hc.order = FALSE, 
           type = "lower",
           lab = TRUE,
           lab_size = 2.5,
           outline.color = "white")
corplot.all


```
```{r across each season}
var.cont <- env.model %>% filter(season %in% c("1_winter")) %>% 
    select(c(NDBI, NDBI_prevsum, #NISI, NDISI, ENDISI, 
             NDVI, NDVI_prevsum, #SAVI, MNDWI, 
             LST, LST_prevsum, 
             ppt_sum, ppt_sum_prevsum,
             #temp_max, temp_max_prevsum, 
             temp_min, temp_min_prevsum,
             # abundance, rich, 
             survey_year
             ))
corplot.win <- ggcorrplot(cor((var.cont), use = "complete.obs", method = "spearman"), 
           hc.order = FALSE, 
           type = "lower",
           lab = TRUE,
           lab_size = 2.5,
           outline.color = "white")
corplot.win


var.cont <- env.model %>% filter(season %in% c("2_spring")) %>% 
    select(c(NDBI, NDBI_prevsum, #NISI, NDISI, ENDISI, 
             NDVI, NDVI_prevsum, #SAVI, MNDWI, 
             LST, LST_prevsum, 
             ppt_sum, ppt_sum_prevsum,
             #temp_max, temp_max_prevsum, 
             temp_min, temp_min_prevsum,
             # abundance, rich, 
             survey_year
             ))
corplot.spr <- ggcorrplot(cor((var.cont), use = "complete.obs", method = "spearman"), 
           hc.order = FALSE, 
           type = "lower",
           lab = TRUE,
           lab_size = 2.5,
           outline.color = "white")
corplot.spr

ggsave("~/GitHub/caplter-birdcommunitydynamics/figures/collinearity/collinearity_envvariables_corebirds_1winter.png",
       corplot.win,
       width = 6,
       height = 4,
       units = "in",
       dpi = 300)

ggsave("~/GitHub/caplter-birdcommunitydynamics/figures/collinearity/collinearity_envvariables_corebirds_2spring.png",
       corplot.spr,
       width = 6,
       height = 4,
       units = "in",
       dpi = 300)
```
### LST vs. Minimum Air Temperature
```{r}

env.model %>% filter(!season %in% c("annual")) %>%
          ggplot(aes(x = LST, y = temp_min)) +
  theme_classic() + 
  geom_point(alpha = 0.3, shape = 16) +
  geom_smooth(method = "lm")+
  stat_cor(method = "spearman", cor.coef.name = "rho", p.accuracy = 0.001, r.accuracy = 0.001, color = "black") +
  labs(x = "LST (°C)", y = "Air Temperature (°C)") 
# coord_cartesian(ylim = c(-1.5, 1.5)) +
# geom_boxplot(col = "green4")



# Within season correlations
(LST.vs.airtemp <- env.model %>% filter(!season %in% c("annual")) %>%
          ggplot(aes(x = LST, y = temp_min, group = season, col = season)) +
  theme_classic() + 
  geom_point(alpha = 0.3, shape = 16) +
  geom_smooth(method = "lm")+
  # scale_color_brewer(palette = "Dark2") +
  scale_color_manual(values = pal_season) +
  stat_cor(method = "spearman", cor.coef.name = "rho", p.accuracy = 0.001, r.accuracy = 0.001, color = "black") +
  labs(x = "LST", y = "Air Temperature") )


ggsave("./figures/collinearity/LST_vs_AirTemp_birdsites_01to16.png",
       LST.vs.airtemp,
       width = 6,
       height = 4,
       units = "in",
       dpi = 300)


```
## Impervious Surface Current vs. Recent
```{r}

# Overall collinearity
env.model %>% filter(!season %in% c("annual")) %>%
          ggplot(aes(x = NDBI, y = NDBI_prevsum)) +
  theme_classic() + 
  geom_point(alpha = 0.3, shape = 16) +
  geom_smooth(method = "lm")+
  scale_color_brewer(palette = "Dark2") +
  stat_cor(method = "spearman", cor.coef.name = "rho", p.accuracy = 0.001, r.accuracy = 0.001, color = "black") +
  labs(x = "NDBI - Current", y = "NDBI - Recent") 
# coord_cartesian(ylim = c(-1.5, 1.5)) +
# geom_boxplot(col = "green4")

# within each season
env.model %>% filter(!season %in% c("annual")) %>%
          ggplot(aes(x = NDBI, y = NDBI_prevsum, group = season, col = season)) +
  theme_classic() + 
  geom_point(alpha = 0.3, shape = 16) +
  geom_smooth(method = "lm", se = FALSE)+
  scale_color_brewer(palette = "Dark2") +
  scale_color_manual(values = pal_season) +
  stat_cor(method = "spearman", cor.coef.name = "rho", p.accuracy = 0.001, r.accuracy = 0.001, color = "black") +
  labs(x = "NDBI - Current", y = "NDBI - Recent")


```

### Impervious Surface vs. Vegetation
One pattern to note here is estimates of NDBI/urbanization (more properly "imperviousness") are very negatively correlated with NDVI, in ways that vary across years and seasons. In part, this is an area with more leafy green vegetation canopy will have the appearance of being less impervious (based on its surface reflectance characteristics)
```{r}

# GLMM of NDBI vs NDVI and year
    # m1 <- glmmTMB(NDBI ~ NDVI_std + year_std + (1|site_code), data = data.env.model)
    # summary(m1)
    # 
    # plot(ggpredict(m1, terms = "year_std"))
    # plot(ggpredict(m1, terms = "NDVI_std"))


# Overall collinearity
env.model %>% filter(!season %in% c("annual")) %>%
          ggplot(aes(x = NDBI, y = NDVI)) +
  theme_classic() + 
  geom_point(alpha = 0.3, shape = 16) +
  geom_smooth(method = "lm")+
  scale_color_brewer(palette = "Dark2") +
  stat_cor(method = "spearman", cor.coef.name = "rho", p.accuracy = 0.001, r.accuracy = 0.001, color = "black") +
  labs(x = "Impervious Surface (NDBI)", y = "Vegetation (NDVI)") 
# coord_cartesian(ylim = c(-1.5, 1.5)) +
# geom_boxplot(col = "green4")

# within each season
(ndbi.vs.ndvi <- env.model %>% filter(!season %in% c("annual")) %>%
          ggplot(aes(x = NDBI, y = NDVI, group = season, col = season)) +
  theme_classic() + 
  geom_point(alpha = 0.3, shape = 16) +
  geom_smooth(method = "lm")+
  # scale_color_brewer(palette = "Dark2") +
  scale_color_manual(values = pal_season) +
  coord_cartesian(ylim = c(0, 0.75)) +
  stat_cor(method = "spearman", cor.coef.name = "rho", p.accuracy = 0.001, r.accuracy = 0.001, color = "black") +
  labs(x = "Impervious Surface (NDBI)", y = "Vegetation (NDVI)") +
  theme(
    legend.position = "none",
    # axis.text.x = element_text(angle = 30, vjust = 1, hjust=1, size = 8),
    # axis.text.y = element_text(size = 8),
    axis.title = element_text(face = "bold", size = 16)
  ))

ggsave("./figures/collinearity/NDBI_vs_NDVI_birdsites_01to16.png",
       ndbi.vs.ndvi,
       width = 6,
       height = 4,
       units = "in",
       dpi = 300)


```
### Impervious Surface vs. LST
```{r}

# Overall collinearity
env.model %>% filter(!season %in% c("annual")) %>%
          ggplot(aes(x = NDBI, y = LST)) +
  theme_classic() + 
  geom_point(alpha = 0.3, shape = 16) +
  geom_smooth(method = "lm")+
  scale_color_brewer(palette = "Dark2") +
  stat_cor(method = "spearman", cor.coef.name = "rho", p.accuracy = 0.001, r.accuracy = 0.001, color = "black") +
  labs(x = "Impervious Surface (NDBI)", y = "Land Surface Temperature (°C)") 
# coord_cartesian(ylim = c(-1.5, 1.5)) +
# geom_boxplot(col = "green4")

# within each season
(ndbi.vs.lst <- env.model %>% filter(!season %in% c("annual")) %>%
          ggplot(aes(x = NDBI, y = LST, group = season, col = season)) +
  theme_classic() + 
  geom_point(alpha = 0.3, shape = 16) +
  geom_smooth(method = "lm")+
  # scale_color_brewer(palette = "Dark2") +
  scale_color_manual(values = pal_season) +
  coord_cartesian(ylim = c(10, 70)) +
  # stat_cor(method = "spearman", cor.coef.name = "rho", p.accuracy = 0.001, r.accuracy = 0.001, color = "black") +
  labs(x = "Impervious Surface (NDBI)", y = "Land Surface Temperature (°C)")+
  theme(
    legend.position = "none",
    # axis.text.x = element_text(angle = 30, vjust = 1, hjust=1, size = 8),
    # axis.text.y = element_text(size = 8),
    axis.title = element_text(face = "bold", size = 16)
  ))

ggsave("./figures/collinearity/NDBI_vs_LST_birdsites_01to16.png",
       ndbi.vs.lst,
       width = 6,
       height = 4,
       units = "in",
       dpi = 300)


```

### Impervious Surface vs. Precipitation
```{r}

# Overall collinearity
env.model %>% filter(!season %in% c("annual")) %>%
          ggplot(aes(x = NDBI, y = ppt_sum)) +
  theme_classic() + 
  geom_point(alpha = 0.3, shape = 16) +
  geom_smooth(method = "lm")+
  scale_color_brewer(palette = "Dark2") +
  stat_cor(method = "spearman", cor.coef.name = "rho", p.accuracy = 0.001, r.accuracy = 0.001, color = "black") +
  labs(x = "Impervious Surface (NDBI)", y = "Precipitation (mm)") 
# coord_cartesian(ylim = c(-1.5, 1.5)) +
# geom_boxplot(col = "green4")

# within each season
(ndbi.vs.ppt <- env.model %>% filter(!season %in% c("annual")) %>%
          ggplot(aes(x = NDBI, y = ppt_sum, group = season, col = season)) +
  theme_classic() + 
  geom_point(alpha = 0.3, shape = 16) +
  geom_smooth(method = "lm")+
  # scale_color_brewer(palette = "Dark2") +
  scale_color_manual(values = pal_season) +
  coord_cartesian(ylim = c(0, 400)) +
  # stat_cor(method = "spearman", cor.coef.name = "rho", p.accuracy = 0.001, r.accuracy = 0.001, color = "black") +
  labs(x = "Impervious Surface (NDBI)", y = "Precipitation (mm)") +
  theme(
    legend.position = "none",
    # axis.text.x = element_text(angle = 30, vjust = 1, hjust=1, size = 8),
    # axis.text.y = element_text(size = 8),
    axis.title = element_text(face = "bold", size = 16)
  ))

ggsave("./figures/collinearity/NDBI_vs_ppt_birdsites_01to16.png",
       ndbi.vs.ppt,
       width = 6,
       height = 4,
       units = "in",
       dpi = 300)


```

### LST vs. NDVI
```{r}
# Across all Seasons
env.model %>% filter(!season %in% c("annual")) %>%
  ggplot(aes(x = NDVI, y = LST)) +
  theme_classic() + 
  geom_point(alpha = 0.3, shape = 16) +
  geom_smooth(method = "lm")+
  scale_color_brewer(palette = "Dark2") +
  stat_cor(method = "spearman", cor.coef.name = "rho", p.accuracy = 0.001, r.accuracy = 0.001, color = "black") +
  labs(x = "Vegetation (NDVI)", y = "Land Surface Temperature (°C)")

# Within each season
(LST.vs.NDVI <- env.model %>%  filter(!season %in% c("annual")) %>%
  ggplot(aes(x = NDVI, y = LST, group = season, col = season)) +
    theme_classic() + 
    geom_point(alpha = 0.3, shape = 16) +
    geom_smooth(method = "lm")+
  # scale_color_brewer(palette = "Dark2") +
  scale_color_manual(values = pal_season) +
  coord_cartesian(ylim = c(10, 70)) +
    # stat_cor(method = "spearman", cor.coef.name = "rho", p.accuracy = 0.001, r.accuracy = 0.001, color = "black") +
    labs(x = "Vegetation (NDVI)", y = "Land Surface Temperature (°C)")+
  theme(
    legend.position = "none",
    # axis.text.x = element_text(angle = 30, vjust = 1, hjust=1, size = 8),
    # axis.text.y = element_text(size = 8),
    axis.title = element_text(face = "bold", size = 16)
    )
)

# coord_cartesian(ylim = c(-1.5, 1.5)) +
# geom_boxplot(col = "green4")


ggsave("./figures/collinearity/LST_vs_NDVI_birdsites_01to16.png",
       LST.vs.NDVI,
       width = 6,
       height = 4,
       units = "in",
       dpi = 300)
```

### LST vs. Precipitation
```{r}
# Across all Seasons
env.model %>% filter(!season %in% c("annual")) %>%
  ggplot(aes(x = ppt_sum, y = LST)) +
  theme_classic() + 
  geom_point(alpha = 0.3, shape = 16) +
  geom_smooth(method = "lm")+
  scale_color_brewer(palette = "Dark2") +
  stat_cor(method = "spearman", cor.coef.name = "rho", p.accuracy = 0.001, r.accuracy = 0.001, color = "black") +
  labs(x = "Precipitation (mm)", y = "Land Surface Temperature (°C)")

# Within each season
(LST.vs.ppt <- env.model %>%  filter(!season %in% c("annual")) %>%
  ggplot(aes(x = ppt_sum, y = LST, group = season, col = season)) +
    theme_classic() + 
    geom_point(alpha = 0.3, shape = 16) +
    geom_smooth(method = "lm")+
    # scale_color_brewer(palette = "Dark2") +
  scale_color_manual(values = pal_season) +
  coord_cartesian(ylim = c(10, 70)) +
    # stat_cor(method = "spearman", cor.coef.name = "rho", p.accuracy = 0.001, r.accuracy = 0.001, color = "black") +
    labs(x = "Precipitation (mm)", y = "Land Surface Temperature (°C)")+
  theme(
    legend.position = "none",
    # axis.text.x = element_text(angle = 30, vjust = 1, hjust=1, size = 8),
    # axis.text.y = element_text(size = 8),
    axis.title = element_text(face = "bold", size = 16)
  )
)

# coord_cartesian(ylim = c(-1.5, 1.5)) +
# geom_boxplot(col = "green4")


ggsave("./figures/collinearity/LST_vs_ppt_birdsites_01to16.png",
       LST.vs.ppt,
       width = 6,
       height = 4,
       units = "in",
       dpi = 300)
```


### Precipitation vs. NDVI
```{r}
# Across all Seasons
env.model %>% ggplot(aes(x = ppt_sum, y = NDVI)) +
  theme_classic() + 
  geom_point(alpha = 0.3, shape = 16) +
  geom_smooth(method = "lm")+
  stat_cor(method = "spearman", cor.coef.name = "rho", p.accuracy = 0.001, r.accuracy = 0.001, color = "black") +
  labs(x = "Precipitation (mm)", y = "Vegetation (NDVI)")

# Within each season
(ppt.vs.ndvi <- env.model %>%  filter(!season %in% c("annual")) %>%
  ggplot(aes(x = ppt_sum, y = NDVI, group = season, col = season)) +
    theme_classic() + 
    geom_point(alpha = 0.3, shape = 16) +
    geom_smooth(method = "lm")+
    # scale_color_brewer(palette = "Dark2") +
  scale_color_manual(values = pal_season) +
  coord_cartesian(ylim = c(0, 0.75)) +
    # stat_cor(method = "spearman", cor.coef.name = "rho", p.accuracy = 0.001, r.accuracy = 0.001, color = "black") +
    labs(x = "Precipitation (mm)", y = "Vegetation (NDVI)")+
  theme(
    legend.position = "none",
    # axis.text.x = element_text(angle = 30, vjust = 1, hjust=1, size = 8),
    # axis.text.y = element_text(size = 8),
    axis.title = element_text(face = "bold", size = 16)
  )
)

# coord_cartesian(ylim = c(-1.5, 1.5)) +
# geom_boxplot(col = "green4")


ggsave("./figures/collinearity/ppt_vs_NDVI_birdsites_01to16.png",
       ppt.vs.ndvi,
       width = 6,
       height = 4,
       units = "in",
       dpi = 300)
```




### Trends in Precipitation and Temperature
```{r}

# ggplot(data.win, aes(x = survey_year - 1, y = LST_monsoon)) +
#   theme_classic() + 
#   geom_point(col = "red4") +
#   geom_smooth(col = "red4")

# winter precipitation
data.env %>% filter(season == "1_winter") %>%
ggplot(aes(x = as.factor(survey_year), y = ppt_sum)) +
  theme_classic() + 
  geom_point(col = "blue4") +
  coord_cartesian(ylim = c(0, 320)) +
  geom_boxplot(col = "blue4")
# spring (early summer) precipitation
data.env %>% filter(season == "2_spring") %>%
ggplot(aes(x = as.factor(survey_year), y = ppt_sum)) +
  theme_classic() + 
  geom_point(col = "blue4") +
  coord_cartesian(ylim = c(0, 320)) +
  geom_boxplot(col = "blue4")

# Late summer (monsoon season) precipitation and temperature
# ggplot(data.win, aes(x = survey_year - 1, y = ppt_sum_monsoon)) +
#   theme_classic() + 
#   geom_point(col = "blue4") +
#   geom_smooth(col = "blue4")
data.env %>% filter(season == "3_summer") %>%
ggplot(aes(x = as.factor(survey_year), y = ppt_sum_prevsum)) +
  theme_classic() + 
  geom_point(col = "blue4") +
  coord_cartesian(ylim = c(0, 320)) +
  geom_boxplot(col = "blue4")
# 

data.env %>% filter(season == "4_fall") %>%
ggplot(aes(x = as.factor(survey_year), y = ppt_sum_prevsum)) +
  theme_classic() + 
  geom_point(col = "blue4") +
  coord_cartesian(ylim = c(0, 320)) +
  geom_boxplot(col = "blue4")

data.env %>% filter(season == "annual") %>%
ggplot(aes(x = as.factor(survey_year), y = ppt_sum_prevsum)) +
  theme_classic() + 
  geom_point(col = "blue4") +
  # coord_cartesian(ylim = c(0, 320)) +
  geom_boxplot(col = "blue4")

data.env %>% filter(!season %in% c("annual")) %>%
ggplot(aes(x = as.factor(survey_year), y = ppt_sum, fill = season)) +
  theme_classic() + 
  # geom_point(col = "blue4") +
  scale_fill_manual(values = pal_season) +
  coord_cartesian(ylim = c(0, 320)) +
  geom_boxplot()
```
# DIFFERENCES AMONG ALL SEASONS AND YEARS

### Test for differences in environmental variables over time
```{r differences among all seasons}
    
# how about among seasons?   
    # among all seasons
    kruskal.test(NDBI ~ season, data = env.model %>% filter(season != "annual"))
    kruskal.test(NDVI ~ season, data = env.model %>% filter(season != "annual"))
    kruskal.test(LST ~ season, data = env.model %>% filter(season != "annual"))
    kruskal.test(ppt_sum ~ season, data = env.model %>% filter(season != "annual"))    
```

```{r differences among all years}
# test for significant variation among years using the Kruskal-Wallis Rank Sum Test
    kruskal.test(NDBI ~ survey_year, data = data.env)
    kruskal.test(NDBI ~ survey_year, data = data.env)
    kruskal.test(NDVI ~ survey_year, data = data.env)
    kruskal.test(NDVI ~ survey_year, data = data.env)
    kruskal.test(LST ~ survey_year, data = data.env)
    kruskal.test(LST ~ survey_year, data = data.env)
    kruskal.test(ppt_sum ~ survey_year, data = data.env)
    kruskal.test(ppt_sum ~ survey_year, data = data.env)
    # looks like they all vary among years
      # p < 0.05 but > 0.01
    
# and how does the variation among years differ among seasons?
    # kruskal.test(NDBI ~ survey_year, data = (data.env %>% filter(season == "1_winter")))
    # kruskal.test(NDBI ~ survey_year, data = (data.env %>% filter(season == "2_spring")))
    # kruskal.test(NDBI ~ survey_year, data = (data.env %>% filter(season == "3_summer")))
    # kruskal.test(NDBI ~ survey_year, data = (data.env %>% filter(season == "4_fall")))
    # # summer NDBI does not vary among years, but it's close (p >0.05 but < 0.1)
    # kruskal.test(NDVI ~ survey_year, data = (data.env %>% filter(season == "1_winter")))
    # kruskal.test(NDVI ~ survey_year, data = (data.env %>% filter(season == "2_spring")))
    # kruskal.test(NDVI ~ survey_year, data = (data.env %>% filter(season == "3_summer")))
    # kruskal.test(NDVI ~ survey_year, data = (data.env %>% filter(season == "4_fall")))
    # # significant for all seasons
    # kruskal.test(LST ~ survey_year, data = (data.env %>% filter(season == "1_winter")))
    # kruskal.test(LST ~ survey_year, data = (data.env %>% filter(season == "2_spring")))
    # kruskal.test(LST ~ survey_year, data = (data.env %>% filter(season == "3_summer")))
    # kruskal.test(LST ~ survey_year, data = (data.env %>% filter(season == "4_fall")))
    # # significant for all seasons
    # kruskal.test(ppt_sum ~ survey_year, data = (data.env %>% filter(season == "1_winter")))
    # kruskal.test(ppt_sum ~ survey_year, data = (data.env %>% filter(season == "2_spring")))
    # kruskal.test(ppt_sum ~ survey_year, data = (data.env %>% filter(season == "3_summer")))
    # kruskal.test(ppt_sum ~ survey_year, data = (data.env %>% filter(season == "4_fall")))

```

```{r between season differences by location type}

    # how about a difference between site types (ESCA vs. riparian)?
      wilcox.test(NDBI ~ loc_type, data = (env.model %>% filter(season == "1_winter")))    # not significant (close)
      wilcox.test(NDBI ~ loc_type, data = (env.model %>% filter(season == "2_spring")))    # not significant
      wilcox.test(NDBI ~ loc_type, data = (env.model %>% filter(season == "3_summer")))    # not significant 
      wilcox.test(NDBI ~ loc_type, data = (env.model %>% filter(season == "4_fall")))      # not significant
      
      wilcox.test(NDVI ~ loc_type, data = (env.model %>% filter(season == "1_winter")))    # significant
      wilcox.test(NDVI ~ loc_type, data = (env.model %>% filter(season == "2_spring")))    # significant
      wilcox.test(NDVI ~ loc_type, data = (env.model %>% filter(season == "3_summer")))    # significant
      wilcox.test(NDVI ~ loc_type, data = (env.model %>% filter(season == "4_fall")))      # significant
      
      wilcox.test(LST ~ loc_type, data = (env.model %>% filter(season == "1_winter")))   # significant
      wilcox.test(LST ~ loc_type, data = (env.model %>% filter(season == "2_spring")))   # significant
      wilcox.test(LST ~ loc_type, data = (env.model %>% filter(season == "3_summer")))   # significant
      wilcox.test(LST ~ loc_type, data = (env.model %>% filter(season == "4_fall")))     # significant
      
      wilcox.test(ppt_sum ~ loc_type, data = (env.model %>% filter(season == "1_winter"))) # not significant
      wilcox.test(ppt_sum ~ loc_type, data = (env.model %>% filter(season == "2_spring"))) # not significant
      wilcox.test(ppt_sum ~ loc_type, data = (env.model %>% filter(season == "3_summer"))) # not significant
      wilcox.test(ppt_sum ~ loc_type, data = (env.model %>% filter(season == "4_fall")))   # not significant
    # So, there are significant differences in NDVI and temperature, which is not too surprising
    # But this does not mean there is necessarily a significant difference in their trends
```


# AMONG-YEAR VARIATION AND TRENDS IN MODELED ENVIRONMENTAL DATA

### NDBI trends

```{r temporal variation in NDBI}
# between winter and spring
    wilcox.test(NDBI ~ season, data = (env.model %>% filter(season %in% c("1_winter", "2_spring"))))

# NDBI: significant difference only during the winter (almost in fall)
    kruskal.test(NDBI ~ survey_year, data = (env.model %>% filter(season == "1_winter")))
    kruskal.test(NDBI ~ survey_year, data = (env.model %>% filter(season == "2_spring")))
    kruskal.test(NDBI ~ survey_year, data = (env.model %>% filter(season == "3_summer")))
    kruskal.test(NDBI ~ survey_year, data = (env.model %>% filter(season == "4_fall")))
    kruskal.test(NDBI ~ survey_year, data = (env.model %>% filter(season == "annual")))
 
```

```{r NDBI trend}
# across all years, NDBI is not increasing at this sites
# in fact, winter, summer, fall and annual NDBI appear to have significant negative correlations with year (decrease over time)
# data.env %>% 
#   ggplot(aes(x = survey_year, y = NDBI, group = season, color = season)) +
#   geom_point(alpha = 0.3, shape = 16) +
#   geom_smooth(method = "lm") +
#   scale_color_brewer(palette = "Dark2") +
#   stat_cor(method = "spearman", cor.coef.name = "rho", p.accuracy = 0.001, r.accuracy = 0.001, color = "black") +
#   labs(x = "Year", y = "Impervious Surface (NDBI)") +
#   theme_bw()   


# for only the bird survey years and sites, NDBI has increased significantly, but not during spring
(plot.trend <- env.model %>% filter(!season %in% c("annual")) %>%
  ggplot(aes(x = survey_year, y = NDBI, group = season, color = season)) +
  geom_point(alpha = 0.3, shape = 16, position = position_dodge(width = 0.6)) +
  geom_smooth(method = "gam", se = FALSE) +
  # scale_color_brewer(palette = "Dark2") +
  scale_color_manual(values = pal_season) +
  stat_cor(method = "spearman", cor.coef.name = "rho", p.accuracy = 0.001, r.accuracy = 0.001, color = "black") +
  # geom_smooth(aes)
  labs(x = "Year", y = "Impervious Surface (NDBI)") +
  theme_bw()+
    theme(
    legend.position = "none",
    axis.text.x = element_text(angle = 30, vjust = 1, hjust=1, size = 16),
    axis.text.y = element_text(size = 16),
    axis.title = element_text(face = "bold", size = 16)
  ))

# lines, points, and error bars
(plot.trend <- env.model %>% filter(!season == "annual") %>% reframe(mean = mean(NDBI), median = median(NDBI),
                                    sd = sd(NDBI), se = sd(NDBI)/sqrt(length(NDBI)), 
                                    .by = c(survey_year, season)) %>%
  ggplot(aes(x = survey_year, y = mean, group = season, color = season)) +
  geom_errorbar(aes(ymin = (mean-1.96*se), ymax = (mean+1.96*se)), width = 0.4, size = 0.4, position = position_dodge(width = 0.5),
                alpha = 0.5) +
  geom_point(aes(shape = season), position = position_dodge(width = 0.5)) +
  scale_shape_manual(values=c(16, 17, 15, 18))+
  geom_line(linewidth = 0.5, position = position_dodge(width = 0.5)) +
  # scale_color_brewer(palette = "Dark2") +
  scale_color_manual(values = pal_season) +
  scale_y_continuous(labels = label_number(accuracy = 0.01)) +
  # stat_cor(method = "spearman", cor.coef.name = "rho", p.accuracy = 0.001, r.accuracy = 0.001, color = "black") +
  # coord_cartesian(ylim = c(0,30)) +
  labs(x = "Year", y = "Impervious\nSurface (NDBI)") +
  theme_bw()+
    theme(
    legend.position = "none",
    axis.text.x = element_text(angle = 30, vjust = 1, hjust=1, size = 7),
    axis.text.y = element_text(size = 7),
    axis.title.x = element_text(vjust = 2),
    axis.title = element_text(face = "bold", size = 8)
  ))

ggsave("./figures/trends_env/env_NDBI_birdsites_01to16.png",
       plot.trend,
       width = 2.5,
       height = 2,
       units = "in",
       dpi = 300)


#
(plot.trend <- env.model %>% filter(!season == "annual") %>% 
    reframe(median = median(NDBI), sd = sd(NDBI), se = sd(NDBI)/length(NDBI), .by = c(survey_year, season)) %>%
  ggplot(aes(x = survey_year, y = median, group = season, color = season)) +
    geom_errorbar(aes(ymin = (median-1.96*se), ymax = (median+1.96*se)), width = 0.1) +
    geom_point(shape = 16) +
    geom_line(linewidth = 1) +
    scale_color_brewer(palette = "Dark2") +
    # stat_cor(method = "spearman", cor.coef.name = "rho", p.accuracy = 0.001, r.accuracy = 0.001, color = "black") +
    # coord_cartesian(ylim = c(0,40)) +
    labs(x = "Year", y = "Impervious Surface (NDBI)") +
    theme_bw()+
      theme(
      legend.position = "none",
      axis.text.x = element_text(angle = 30, vjust = 1, hjust=1, size = 14),
      axis.text.y = element_text(size = 14),
      axis.title = element_text(face = "bold", size = 16)
    ))


(plot.trend <- env.model %>% filter(!season == "annual") %>% #filter(season %in% c("1_winter", "2_spring")) %>%
  ggplot(aes(x = as.factor(survey_year), y = NDBI, fill = season)) +
  theme_classic() + 
  # geom_point(col = "blue4") +
  # coord_cartesian(ylim = c(0, 320)) +
  geom_boxplot(color = "grey40", outlier.alpha = 0.2, linewidth = 0.3)+
  # scale_fill_brewer(palette = "Dark2") +
  scale_fill_manual(values = pal_season) +
  labs(x = "Year", y = "Impervious Surface (NDBI)") +
  theme(
    legend.position = "none",
    axis.text.x = element_text(angle = 30, vjust = 1, hjust=1, size = 14),
    axis.text.y = element_text(size = 14),
    axis.title = element_text(face = "bold", size = 16)
  ))




glmmTMB(NDBI ~ year_std  
                  + (1|loc_type)
                  # + (1 + year_std|site_code)
                  , data = (env.model %>% filter(season == "1_winter"))) %>% summary()




# We can also test these relationships using GLMMs with site as a random intercept
  m1 <- glmmTMB(NDBI ~ year_std + (1|site_code) , data = data.win)
  m2 <- glmmTMB(NDBI ~ year_std + (1|site_code) , data = data.spr)

  print("Winter")
  # summary(m1)
  # check_overdispersion(m1)   # no significant overdispersion
  round(confint(m1), 3)
  # ranef(m1)
  # r.squaredGLMM(m1)
  
  print("Spring")
  # summary(m2)
  # check_overdispersion(m2)   # no significant overdispersion
  round(confint(m2), 3)
  # ranef(m2)
  # r.squaredGLMM(m2)
```

```{r}

# differences among location types
    env.model %>% filter(season %in% c("1_winter")) %>%
      ggplot(aes(x = survey_year, y = NDBI, group = loc_type, color = loc_type)) +
      geom_point(alpha = 0.3, shape = 16) +
      geom_smooth(method = "gam") +
      scale_color_brewer(palette = "Dark2") +
      stat_cor(method = "spearman", cor.coef.name = "rho", p.accuracy = 0.001, r.accuracy = 0.001, color = "black") +
      # geom_smooth(aes)
      labs(x = "Year", y = "Impervious Surface (NDBI)") +
      theme_bw()
    
    env.model %>% filter(season %in% c("2_spring")) %>%
      ggplot(aes(x = survey_year, y = NDBI, group = loc_type, color = loc_type)) +
      geom_point(alpha = 0.3, shape = 16) +
      geom_smooth(method = "lm") +
      scale_color_brewer(palette = "Dark2") +
      stat_cor(method = "spearman", cor.coef.name = "rho", p.accuracy = 0.001, r.accuracy = 0.001, color = "black") +
      # geom_smooth(aes)
      labs(x = "Year", y = "Impervious Surface (NDBI)") +
      theme_bw()
    
# By site
    env.model %>% filter(season %in% c("2_spring")) %>%
      ggplot(aes(x = survey_year, y = NDBI, group = site_code, color = site_code)) +
      geom_point(alpha = 0.3, shape = 16) +
      geom_smooth(method = "lm", se = FALSE) +
      # scale_color_brewer(palette = "Dark2") +
      # stat_cor(method = "spearman", cor.coef.name = "rho", p.accuracy = 0.001, r.accuracy = 0.001, color = "black") +
      # geom_smooth(aes)
      gghighlight(loc_type == "riparian") +
      labs(x = "Year", y = "Impervious Surface (NDBI)") +
      theme_bw() +
      theme(
        legend.position = "none"
      )
```

### NDVI Trends

```{r temporal variation NDVI}
# between winter and spring
    wilcox.test(NDVI ~ season, data = (env.model %>% filter(season %in% c("1_winter", "2_spring"))))
    
# NDVI: significant for all seasons
    kruskal.test(NDVI ~ survey_year, data = (env.model %>% filter(season == "1_winter")))
    kruskal.test(NDVI ~ survey_year, data = (env.model %>% filter(season == "2_spring")))
    kruskal.test(NDVI ~ survey_year, data = (env.model %>% filter(season == "3_summer")))
    kruskal.test(NDVI ~ survey_year, data = (env.model %>% filter(season == "4_fall")))
```


```{r NDVI trends}
# for only the bird survey years, NDVI is increasing significantly during spring, but not quite during winter
(plot.trend <- env.model %>% filter(!season == "annual") %>% #filter(season %in% c("1_winter", "2_spring")) %>%
  ggplot(aes(x = survey_year, y = NDVI, group = season, color = season)) +
  geom_point(alpha = 0.3, shape = 16, position = position_dodge(width = 0.6)) +
  geom_smooth(method = "gam", se = FALSE) +
  # scale_color_brewer(palette = "Dark2") +
  scale_color_manual(values = pal_season) +
  # coord_cartesian(ylim = c(0.05, 0.45)) +
  # stat_cor(method = "spearman", cor.coef.name = "rho", p.accuracy = 0.001, r.accuracy = 0.001, color = "black") +
  labs(x = "Year", y = "\nVegetation (NDVI)") +
  theme_bw()+
    theme(
    legend.position = "none",
    axis.text.x = element_text(angle = 30, vjust = 1, hjust=1, size = 16),
    axis.text.y = element_text(size = 16),
    axis.title = element_text(face = "bold", size = 16)
  ))


# lines, points, and error bars
(plot.trend <- env.model %>% filter(!season == "annual") %>% reframe(mean = mean(NDVI), median = median(NDVI),
                                    sd = sd(NDVI), se = sd(NDVI)/sqrt(length(NDVI)), 
                                    .by = c(survey_year, season)) %>%
  ggplot(aes(x = survey_year, y = mean, group = season, color = season)) +
  geom_errorbar(aes(ymin = (mean-1.96*se), ymax = (mean+1.96*se)), width = 0.4, size = 0.4, position = position_dodge(width = 0.5),
                alpha = 0.5) +
  geom_point(aes(shape = season), position = position_dodge(width = 0.5)) +
  scale_shape_manual(values=c(16, 17, 15, 18))+
  geom_line(linewidth = 0.5, position = position_dodge(width = 0.5)) +
  # scale_color_brewer(palette = "Dark2") +
  scale_color_manual(values = pal_season) +
  scale_y_continuous(labels = label_number(accuracy = 0.01)) +
  # stat_cor(method = "spearman", cor.coef.name = "rho", p.accuracy = 0.001, r.accuracy = 0.001, color = "black") +
  # coord_cartesian(ylim = c(0,30)) +
  labs(x = "Year", y = "\nVegetation (NDVI)") +
  theme_bw()+
    theme(
    legend.position = "none",
    axis.text.x = element_text(angle = 30, vjust = 1, hjust=1, size = 7),
    axis.text.y = element_text(size = 7),
    axis.title.x = element_text(vjust = 2),
    axis.title = element_text(face = "bold", size = 8)
  ))

ggsave("./figures/trends_env/env_NDVI_birdsites_01to16.png",
       plot.trend,
       width = 2.5,
       height = 2,
       units = "in",
       dpi = 300)

(plot.trend <- env.model %>% filter(!season == "annual") %>% #filter(season %in% c("1_winter", "2_spring")) %>%
  ggplot(aes(x = as.factor(survey_year), y = NDVI, fill = season)) +
    theme_classic() + 
    # geom_point(col = "blue4") +
    # coord_cartesian(ylim = c(0, 320)) +
    geom_boxplot(color = "grey40", outlier.alpha = 0.2, linewidth = 0.3)+
  # scale_fill_brewer(palette = "Dark2") +
  scale_fill_manual(values = pal_season) +
    labs(x = "Year", y = "Vegetation (NDVI)") +
    theme(
    legend.position = "none",
    axis.text.x = element_text(angle = 30, vjust = 1, hjust=1, size = 14),
    axis.text.y = element_text(size = 14),
    axis.title = element_text(face = "bold", size = 16)
  ))



# GLMMs 
  # m1 <- glmmTMB(NDVI ~ year_std + (1|site_code) , data = data.win)
  # m2 <- glmmTMB(NDVI ~ year_std + (1|site_code) , data = data.spr)
  # 
  # print("Winter")
  # summary(m1)
  # check_overdispersion(m1)   # no significant overdispersion
  # round(confint(m1), 3)
  # ranef(m1)
  # r.squaredGLMM(m1)
  
  # print("Spring")
  # summary(m2)
  # check_overdispersion(m2)   # no significant overdispersion
  # round(confint(m2), 3)
  # ranef(m2)
  # r.squaredGLMM(m2)
```

### LST Trends
```{r temporal variation LST}

    # between winter and spring
    wilcox.test(LST ~ season, data = (env.model %>% filter(season %in% c("1_winter", "2_spring"))))
    
    # LST: significant for all seasons
    kruskal.test(LST ~ survey_year, data = (env.model %>% filter(season == "1_winter")))
    kruskal.test(LST ~ survey_year, data = (env.model %>% filter(season == "2_spring")))
    kruskal.test(LST ~ survey_year, data = (env.model %>% filter(season == "3_summer")))
    kruskal.test(LST ~ survey_year, data = (env.model %>% filter(season == "4_fall")))
    
```

```{r temperature trends}
# across all years, temperature in each season is increasing across sites
# data.env %>% 
#   ggplot(aes(x = survey_year, y = LST, group = season, color = season)) +
#   geom_point(alpha = 0.3, shape = 16) +
#   geom_smooth(method = "lm") +
#   scale_color_brewer(palette = "Dark2") +
#   stat_cor(method = "spearman", cor.coef.name = "rho", p.accuracy = 0.001, r.accuracy = 0.001, color = "black") +
#   labs(x = "Year", y = "Temperature") +
#   theme_bw()

# for only the bird survey years, temperature is increasing significantly during winter, but not quite during spring
(plot.trend <- env.model %>% filter(!season == "annual") %>% #filter(season %in% c("1_winter", "2_spring")) %>%
  ggplot(aes(x = survey_year, y = LST, group = season, color = season)) +
  geom_point(alpha = 0.3, shape = 16, position = position_dodge(width = 0.6)) +
  geom_smooth(method = "gam", se = FALSE) +
  # scale_color_brewer(palette = "Dark2") +
  scale_color_manual(values = pal_season) +
  # stat_cor(method = "spearman", cor.coef.name = "rho", p.accuracy = 0.001, r.accuracy = 0.001, color = "black") +
  labs(x = "Year", y = "Land Surface\nTemperature (°C)") +
  theme_bw()+
    theme(
    legend.position = "none",
    axis.text.x = element_text(angle = 30, vjust = 1, hjust=1, size = 16),
    axis.text.y = element_text(size = 16),
    axis.title = element_text(face = "bold", size = 16)
  ))


# lines, points, and error bars
(plot.trend <- env.model %>% filter(!season == "annual") %>% reframe(mean = mean(LST), median = median(LST),
                                    sd = sd(LST), se = sd(LST)/sqrt(length(LST)), 
                                    .by = c(survey_year, season)) %>%
  ggplot(aes(x = survey_year, y = mean, group = season, color = season)) +
  geom_errorbar(aes(ymin = (mean-1.96*se), ymax = (mean+1.96*se)), width = 0.4, size = 0.4, position = position_dodge(width = 0.5),
                alpha = 0.5) +
  geom_point(aes(shape = season), position = position_dodge(width = 0.5)) +
  scale_shape_manual(values=c(16, 17, 15, 18))+
  geom_line(linewidth = 0.5, position = position_dodge(width = 0.5)) +
  # scale_color_brewer(palette = "Dark2") +
  scale_color_manual(values = pal_season) +
  scale_y_continuous(labels = label_number(accuracy = 0.1)) +
  # stat_cor(method = "spearman", cor.coef.name = "rho", p.accuracy = 0.001, r.accuracy = 0.001, color = "black") +
  # coord_cartesian(ylim = c(0,30)) +
  labs(x = "Year", y = "Land Surface\nTemperature (°C)") +
  theme_bw()+
    theme(
    legend.position = "none",
    axis.text.x = element_text(angle = 30, vjust = 1, hjust=1, size = 7),
    axis.text.y = element_text(size = 7),
    axis.title.x = element_text(vjust = 2),
    axis.title = element_text(face = "bold", size = 8)
  ))

ggsave("./figures/trends_env/env_LST_birdsites_01to16.png",
       plot.trend,
       width = 2.5,
       height = 2,
       units = "in",
       dpi = 300)

(plot.trend <- env.model %>% filter(!season == "annual") %>% #filter(season %in% c("1_winter", "2_spring")) %>%
  ggplot(aes(x = as.factor(survey_year), y = LST, fill = season)) +
    theme_classic() + 
    # geom_point(col = "blue4") +
    # coord_cartesian(ylim = c(0, 320)) +
    geom_boxplot(color = "grey40", outlier.alpha = 0.2, linewidth = 0.3)+
  # scale_fill_brewer(palette = "Dark2") +
  scale_fill_manual(values = pal_season) +
    labs(x = "Year", y = "Land Surface Temperature (°C)") +
    theme(
    legend.position = "none",
    axis.text.x = element_text(angle = 30, vjust = 1, hjust=1, size = 14),
    axis.text.y = element_text(size = 14),
    axis.title = element_text(face = "bold", size = 16)
  ))

# 
# glmmTMB(LST ~ survey_year 
#         + as.factor(season) 
#         # + survey_year*as.factor(season)
#         + (1|site_code) , data = data.env.model) %>% summary()



# GLMMs 
  # m1 <- glmmTMB(LST ~ year_std + (1|site_code) , data = data.win)
  # m2 <- glmmTMB(LST ~ year_std + (1|site_code) , data = data.spr)
  # 
  # print("Winter")
  # summary(m1)
  # check_overdispersion(m1)   # no significant overdispersion
  # round(confint(m1), 3)
  # ranef(m1)
  # r.squaredGLMM(m1)
  
  # print("Spring")
  # summary(m2)
  # check_overdispersion(m2)   # no significant overdispersion
  # round(confint(m2), 3)
  # ranef(m2)
  # r.squaredGLMM(m2)
```

### Precipitation Trends
```{r temporal variation PPT}
    # between winter and spring
    wilcox.test(ppt_sum ~ season, data = (env.model %>% filter(season %in% c("1_winter", "2_spring"))))
    
    # PPT: significant for all seasons
    kruskal.test(ppt_sum ~ survey_year, data = (env.model %>% filter(season == "1_winter")))
    kruskal.test(ppt_sum ~ survey_year, data = (env.model %>% filter(season == "2_spring")))
    kruskal.test(ppt_sum ~ survey_year, data = (env.model %>% filter(season == "3_summer")))
    kruskal.test(ppt_sum ~ survey_year, data = (env.model %>% filter(season == "4_fall")))
```

```{r precipitation trends}
# across all years, 
# precipitation in is generally decreasing in the winter and spring
# and increasing in the summer and fall
# data.env %>% 
#   ggplot(aes(x = survey_year, y = ppt_sum, group = season, col = season, color = season)) +
#   geom_point(alpha = 0.6) +
#   geom_smooth(method = "lm") +
#   scale_color_brewer(palette = "Dark2") +
#   stat_cor(method = "spearman", cor.coef.name = "rho", p.accuracy = 0.001, r.accuracy = 0.001, color = "black") +
#   labs(x = "Year", y = "Precipitation") +
#   theme_bw()

# for only the bird survey years, precipitation is decreasing significantly during winter, but not during spring
(plot.trend <- env.model  %>% filter(!season == "annual") %>% #filter(season %in% c("1_winter", "2_spring")) %>%
  ggplot(aes(x = survey_year, y = ppt_sum, group = season, col = season, color = season)) +
  geom_point(alpha = 0.3, shape = 16, position = position_dodge(width = 0.6)) +
  geom_smooth(method = "gam", se = FALSE) +
  # scale_color_brewer(palette = "Dark2") +
  scale_color_manual(values = pal_season) +
  # stat_cor(method = "spearman", cor.coef.name = "rho", p.accuracy = 0.001, r.accuracy = 0.001, color = "black") +
  labs(x = "Year", y = "Precipitation (mm)") +
  theme_bw()+
    theme(
    legend.position = "none",
    axis.text.x = element_text(angle = 30, vjust = 1, hjust=1, size = 16),
    axis.text.y = element_text(size = 16),
    axis.title = element_text(face = "bold", size = 16)
  ))


# lines, points, and error bars
(plot.trend <- env.model %>% filter(!season == "annual") %>% reframe(mean = mean(ppt_sum), median = median(ppt_sum),
                                    sd = sd(ppt_sum), se = sd(ppt_sum)/sqrt(length(ppt_sum)), 
                                    .by = c(survey_year, season)) %>%
  ggplot(aes(x = survey_year, y = mean, group = season, color = season)) +
  geom_errorbar(aes(ymin = (mean-1.96*se), ymax = (mean+1.96*se)), width = 0.4, size = 0.4, position = position_dodge(width = 0.5),
                alpha = 0.5) +
  geom_point(aes(shape = season), position = position_dodge(width = 0.5)) +
  scale_shape_manual(values=c(16, 17, 15, 18))+
  geom_line(linewidth = 0.5, position = position_dodge(width = 0.5)) +
  # scale_color_brewer(palette = "Dark2") +
  scale_color_manual(values = pal_season) +
  # stat_cor(method = "spearman", cor.coef.name = "rho", p.accuracy = 0.001, r.accuracy = 0.001, color = "black") +
  # coord_cartesian(ylim = c(0,30)) +
  labs(x = "Year", y = "\nPrecipitation (mm)") +
  theme_bw()+
    theme(
    legend.position = "none",
    axis.text.x = element_text(angle = 30, vjust = 1, hjust=1, size = 7),
    axis.text.y = element_text(size = 7),
    axis.title.x = element_text(vjust = 2),
    axis.title = element_text(face = "bold", size = 8)
  ))

ggsave("./figures/trends_env/env_PPT_birdsites_01to16.png",
       plot.trend,
       width = 2.5,
       height = 2,
       units = "in",
       dpi = 300)

(plot.trend <- env.model %>% filter(!season == "annual") %>% #filter(season %in% c("1_winter", "2_spring")) %>%
  ggplot(aes(x = as.factor(survey_year), y = ppt_sum, fill = season)) +
    theme_classic() + 
    # geom_point(col = "blue4") +
    # coord_cartesian(ylim = c(0, 320)) +
    geom_boxplot(color = "grey40", outlier.alpha = 0.2, linewidth = 0.3)+
  # scale_fill_brewer(palette = "Dark2") +
  scale_fill_manual(values = pal_season) +
    labs(x = "Year", y = "Precipitation (mm)") +
    theme(
    legend.position = "none",
    axis.text.x = element_text(angle = 30, vjust = 1, hjust=1, size = 14),
    axis.text.y = element_text(size = 14),
    axis.title = element_text(face = "bold", size = 16)
  )
  )


# GLMMs 
  m1 <- glmmTMB(ppt_sum ~ year_std + (1|site_code) , data = data.win)
  m2 <- glmmTMB(ppt_sum ~ year_std + (1|site_code) , data = data.spr)

  print("Winter")
  # summary(m1)
  # check_overdispersion(m1)   # no significant overdispersion
  round(confint(m1), 3)
  # ranef(m1)
  # r.squaredGLMM(m1)
  
  print("Spring")
  # summary(m2)
  # check_overdispersion(m2)   # no significant overdispersion
  round(confint(m2), 3)
  # ranef(m2)
  # r.squaredGLMM(m2)
```





# BONUS
