==> code/2_2_AvianCommunityTrends_caplter-birdcommunitydynamics.Rmd <==
---
title: "CAP Bird Community Dynamics - Dynamic Community Occupancy Model"
author: "Jeffrey Haight et al."
date: "`r Sys.Date()`"
output:
  word_document: default
  html_document: default
---

Code for testing and visualizing spatial-temporal patterns overall bird community characteristics (abundance and species richness) and functional group abundances

```{r setup, set.seed(54321), include=FALSE}
knitr::opts_chunk$set(
  echo = FALSE,
  warning = FALSE, 
  include = TRUE,
  cache = TRUE
  )
rm(list = ls())
gc()
```

```{r packages, message = FALSE, echo = TRUE, warning = FALSE}
  library(tidyverse)
  library(beepr)  # for notifying you when some code is  done running

# for statistical modeling
  library(performance)  # for checking model performance (multicollinearity, VIF, overdispersion, etc.)
  library(insight)      # for extracting a lot of useful model information, using functions like 'get_variance()'
  library(lme4)
  library(glmmTMB)
  library(lmerTest)     # for getting p-values out of the lmer models
  library(MuMIn)        # for calculating R^2, among other things
  library(vegan)        # for multivariate community analyses
  library(jagsUI)       # for running Bayesian models in JAGS
  library(mgcv)         # for GAM(M)s
  library(itsadug)      # for GAMM plotting

# for plotting
# library(ggplot2)      # should already be in the 'tidyverse'
  library(ggpubr)       # for adding correlations and p-values to ggplot
  library(ggeffects)    # for plotting glmm effects in ggplot
  library(ggcorrplot)   # for correlation plots
  library(GGally)       # similar to ggcorrplot
  library(gghighlight)  # for highlighting groups in ggplot
  library(RColorBrewer)
  library(scales)
  library(lterpalettefinder)
```
```{r color palettes} 

  # import some color palettes to use for visualizing seasons
  load("~/GitHub/caplter-birdcommunitydynamics/data/arizonapalettes.RData")
  # Cool-Wet: color = 6, fill = 1
  # Hot-Dry spring: color = 2, fill = 3
  # Warm-Wet: color = 4, fill = 5
  # Warm-Dry
palette_ggdemo(asu)
palette_ggdemo(wildflower)
palette_ggdemo(biocrust)
palette_ggdemo(sonoran)

# Palettes for the four seasons based on local colors
(pal_season <- c(
  "#2f3ca3", # arroyo lupine
  # "#4ab7c4",
  # "#7291fb",
  # "#70a112", # moss green
  # "#c8ff08", # lichen green (too bright)
  "#78be20", 
  "#d10073",  # prickly pear pink #1
  # "#ff8701", # poppy orange
  "#ffc901", # palo verde yellow
  "#000000"
)) %>% palette_ggdemo()

(pal_season_asu <- c(
  "#00a3e0", 
  "#78be20", 
  "#e74973", 
  "#ff7f32",
  "#000000"
)) %>% palette_ggdemo()

# (pal_season_light <- c(
#                   "#7291fb",
#                   "#f2ff0d",  # brittlebush yellow
#                   "#ff23c1"   #prickly pear pink #2
#                   )) %>% palette_ggdemo()


```

# Import Data
```{r, include = FALSE}
list.files("~/GitHub/caplter-birdcommunitydynamics/data")

# This contains data assembled by the '2_1_StatisticalModelingSetup_CAPbirds' file
load("~/GitHub/caplter-birdcommunitydynamics/data/modelinputs_CAPbirds2024.RData")


data.win$site_code <- as.factor(data.spr$site_code)
data.spr$site_code <- as.factor(data.spr$site_code)
```
# COMMUNITY-LEVEL ANALYSIS
### Differences between seasons
```{r differences between seasons}
# test for significant differences between seasons using the Wilcoxan Rank Sum
    wilcox.test(abundance ~ season, data = data.div)
    # wilcox.test(abund_weight_con ~ season, data = data.div)
    wilcox.test(rich ~ season, data = data.div)
    # wilcox.test(rich_weight_con ~ season, data = data.div)
    # wilcox.test(beta_s_mean ~ season, data = data.div)
    # there are significant differences between seasons, though abundance is not too significantly different
    
    length(spp.known)
```

### Variation among Years
```{r variation among years - community}
# test for significant variation among years using the Kruskal-Wallis Rank Sum Test
    kruskal.test(abundance ~ survey_year, data = data.win)
    # kruskal.test(abund_weight_con ~ survey_year, data = data.win)
    kruskal.test(abundance ~ survey_year, data = data.spr)
    # kruskal.test(abund_weight_con ~ survey_year, data = data.spr)
    kruskal.test(rich ~ survey_year, data = data.win)
    # kruskal.test(rich_weight_con ~ survey_year, data = data.win)
    kruskal.test(rich ~ survey_year, data = data.spr)
    # kruskal.test(rich_weight_con ~ survey_year, data = data.spr)
    # kruskal.test(beta_s_mean ~ survey_year, data = data.win)
    # kruskal.test(beta_s_mean ~ survey_year, data = data.spr)
    # and yes, there is very significant variation in the overall bird community over the years

```

# GUILD-LEVEL ANALYSIS
### Variation between seasons
First, we may wonder if there is a seasonal difference in guild abundance. This can also be examined by comparing the intercepts between the two seasonal models (e.g. significant difference if the 95% CIs don't overlap, or if they don't overlap the mean)
```{r seed dispersers - seasonal diff}
wilcox.test(abund_SeedDisp ~ season, data = data.div) #%>% str()
# significant difference between seasons 
```

```{r Insectivores - seasonal diff}
wilcox.test(abund_Invert ~ season, data = data.div)
# no significant difference between seasons
```
```{r pollinators - seasonal diff}
wilcox.test(abund_pollen ~ season, data = data.div)
# significant difference between seasons
```
### Variation among years
```{r variation among years - guilds}
# test for significant variation among years using the Kruskal-Wallis Rank Sum Test
    kruskal.test(abund_SeedDisp ~ survey_year, data = data.win)
    kruskal.test(abund_SeedDisp ~ survey_year, data = data.spr)
    
    kruskal.test(abund_Invert ~ survey_year, data = data.win)
    kruskal.test(abund_Invert ~ survey_year, data = data.spr)
    
    kruskal.test(abund_pollen ~ survey_year, data = data.win)
    kruskal.test(abund_pollen ~ survey_year, data = data.spr)

```



# FIGURE 2: Trends in total bird abundance, species richness, and guild abundance
```{r abundance trend}
(plot.trend <- data.div %>% 
  ggplot(aes(x = survey_year, y = abundance, group = season, color = season)) +
  geom_point(alpha = 0.3, shape = 16, position = position_dodge(width = 0.6)) +
  geom_smooth(method = "lm") +
  # scale_color_brewer(palette = "Dark2") +
  scale_color_manual(values = pal_season) +
  stat_cor(method = "spearman", cor.coef.name = "rho", p.accuracy = 0.001, r.accuracy = 0.001, color = "black") +
  coord_cartesian(ylim = c(0,100)) +
  labs(x = "Year", y = "Total Bird Abundance \n(no./survey)") +
  theme_bw()+
    theme(
    legend.position = "none",
    axis.text.x = element_text(angle = 30, vjust = 1, hjust=1, size = 12),
    axis.text.y = element_text(size = 14),
    axis.title = element_text(face = "bold", size = 16)
  ))

# boxplot
(plot.trend <- data.div %>% 
  ggplot(aes(x = as.factor(survey_year), y = abundance, fill = season)) +
    theme_bw() + 
    # geom_point(col = "blue4") +
    geom_boxplot(color = "grey40", outlier.alpha = 0.2, linewidth = 0.3)+
  # scale_fill_brewer(palette = "Dark2") +
  scale_fill_manual(values = pal_season) +
  coord_cartesian(ylim = c(0,65)) +
  labs(x = "Year", y = "Total Bird Abundance \n(no./survey)") +
  theme(
    legend.position = "none",
    axis.text.x = element_text(angle = 40, vjust = 1, hjust=1, size = 14),
    axis.text.y = element_text(size = 14),
    axis.title = element_text(face = "bold", size = 16)
  ))

# lines, points, and error bars
(plot.trend <- data.div %>% reframe(mean = mean(abundance), median = median(abundance),
                                    sd = sd(abundance), se = sd(abundance)/sqrt(length(abundance)), 
                                    .by = c(survey_year, season)) %>%
  ggplot(aes(x = survey_year, y = mean, group = season, color = season)) +
  geom_errorbar(aes(ymin = (mean-1.96*se), ymax = (mean+1.96*se)), width = 0.4, size = 0.4, 
                position = position_dodge(width = 0.4), alpha = 0.5) +
  geom_point(shape = 16, position = position_dodge(width = 0.4)) +
  geom_line(linewidth = 0.5, position = position_dodge(width = 0.4)) +
  # scale_color_brewer(palette = "Dark2") +
  scale_color_manual(values = pal_season) +
  # stat_cor(method = "spearman", cor.coef.name = "rho", p.accuracy = 0.001, r.accuracy = 0.001, color = "black") +
  coord_cartesian(ylim = c(0,40)) +
  labs(x = "Year", y = "Total Bird Abundance \n(no./survey)") +
  theme_bw()+
    theme(
    legend.position = "none",
    axis.text.x = element_text(angle = 30, vjust = 1, hjust=1, size = 14),
    axis.text.y = element_text(size = 14),
    axis.title = element_text(face = "bold", size = 16)
  ))

ggsave("~/GitHub/caplter-birdcommunitydynamics/figures/trends_birds/trend_birds_abundance.png",
       plot.trend,
       width = 6,
       height = 4,
       units = "in",
       dpi = 300)
```
```{r richness trend}
(plot.trend <- data.div %>% 
  ggplot(aes(x = survey_year, y = rich, group = season, color = season)) +
  geom_point(alpha = 0.3, shape = 16, position = position_dodge(width = 0.6)) +
  geom_smooth(method = "lm") +
  # scale_color_brewer(palette = "Dark2") +
  scale_color_manual(values = pal_season) +
  # stat_cor(method = "spearman", cor.coef.name = "rho", p.accuracy = 0.001, r.accuracy = 0.001, color = "black") +
  # coord_cartesian(ylim = c(0,100)) +
  labs(x = "Year", y = "Species Richness \n(no./survey)") +
  theme_bw()+
    theme(
    legend.position = "none",
    axis.text.x = element_text(angle = 30, vjust = 1, hjust=1, size = 14),
    axis.text.y = element_text(size = 14),
    axis.title = element_text(face = "bold", size = 16)
  ))


# boxplot
(plot.trend <- data.div %>% 
  ggplot(aes(x = as.factor(survey_year), y = rich, fill = season)) +
    theme_bw() + 
    # geom_point(col = "blue4") +
    geom_boxplot(color = "grey40", outlier.alpha = 0.2, linewidth = 0.3)+
  # scale_fill_brewer(palette = "Dark2") +
  scale_fill_manual(values = pal_season) +
  coord_cartesian(ylim = c(0,30)) +
  labs(x = "Year", y = "Species Richness \n(no./survey)") +
  theme(
    legend.position = "none",
    axis.text.x = element_text(angle = 40, vjust = 1, hjust=1, size = 14),
    axis.text.y = element_text(size = 14),
    axis.title = element_text(face = "bold", size = 16)
  ))

# lines, points, and error bars
(plot.trend <- data.div %>% reframe(mean = mean(rich), median = median(rich),
                                    sd = sd(rich), se = sd(rich)/sqrt(length(rich)), 
                                    .by = c(survey_year, season)) %>%
  ggplot(aes(x = survey_year, y = mean, group = season, color = season)) +
  geom_errorbar(aes(ymin = (mean-1.96*se), ymax = (mean+1.96*se)), width = 0.4, size = 0.4, 
                position = position_dodge(width = 0.4), alpha = 0.5) +
  geom_point(shape = 16, position = position_dodge(width = 0.4)) +
  geom_line(linewidth = 0.5, position = position_dodge(width = 0.4)) +
  # scale_color_brewer(palette = "Dark2") +
  scale_color_manual(values = pal_season) +
  # stat_cor(method = "spearman", cor.coef.name = "rho", p.accuracy = 0.001, r.accuracy = 0.001, color = "black") +
  coord_cartesian(ylim = c(0,20)) +
  labs(x = "Year", y = "Species Richness \n(no./survey)") +
  theme_bw()+
    theme(
    legend.position = "none",
    axis.text.x = element_text(angle = 30, vjust = 1, hjust=1, size = 14),
    axis.text.y = element_text(size = 14),
    axis.title = element_text(face = "bold", size = 16)
  ))


ggsave("~/GitHub/caplter-birdcommunitydynamics/figures/trends_birds/trend_birds_richness.png",
       plot.trend,
       width = 6,
       height = 4,
       units = "in",
       dpi = 300)
```

```{r seed disperser trend}
(plot.trend <- data.div %>% 
  ggplot(aes(x = survey_year, y = abund_SeedDisp, group = season, color = season)) +
  geom_point(alpha = 0.3, shape = 16, position = position_dodge(width = 0.6)) +
  geom_smooth(method = "lm") +
  # scale_color_brewer(palette = "Dark2") +
  scale_color_manual(values = pal_season) +
  # stat_cor(method = "spearman", cor.coef.name = "rho", p.accuracy = 0.001, r.accuracy = 0.001, color = "black") +
  coord_cartesian(ylim = c(0,40)) +
  labs(x = "Year", y = "Seed Disperser Abundance \n(no./survey)") +
  theme_bw()+
    theme(
    legend.position = "none",
    axis.text.x = element_text(angle = 30, vjust = 1, hjust=1, size = 14),
    axis.text.y = element_text(size = 14),
    axis.title = element_text(face = "bold", size = 16)
  ))


# boxplot
(plot.trend <- data.div %>% 
  ggplot(aes(x = as.factor(survey_year), y = abund_SeedDisp, fill = season)) +
    theme_bw() + 
    # geom_point(col = "blue4") +
    geom_boxplot(color = "grey40", outlier.alpha = 0.2, linewidth = 0.3)+
  # scale_fill_brewer(palette = "Dark2") +
  scale_fill_manual(values = pal_season) +
  coord_cartesian(ylim = c(0,45)) +
  labs(x = "Year", y = "Seed Disperser Abundance \n(no./survey)") +
  theme(
    legend.position = "none",
    axis.text.x = element_text(angle = 40, vjust = 1, hjust=1, size = 14),
    axis.text.y = element_text(size = 14),
    axis.title = element_text(face = "bold", size = 16)
  ))


# lines, points, and error bars
(plot.trend <- data.div %>% reframe(mean = mean(abund_SeedDisp), median = median(abund_SeedDisp),
                                    sd = sd(abund_SeedDisp), se = sd(abund_SeedDisp)/sqrt(length(abund_SeedDisp)), 
                                    .by = c(survey_year, season)) %>%
  ggplot(aes(x = survey_year, y = mean, group = season, color = season)) +
  geom_errorbar(aes(ymin = (mean-1.96*se), ymax = (mean+1.96*se)), width = 0.4, size = 0.4, 
                position = position_dodge(width = 0.4), alpha = 0.5) +
  geom_point(shape = 16, position = position_dodge(width = 0.4)) +
  geom_line(linewidth = 0.5, position = position_dodge(width = 0.4)) +
  # scale_color_brewer(palette = "Dark2") +
  scale_color_manual(values = pal_season) +
  # stat_cor(method = "spearman", cor.coef.name = "rho", p.accuracy = 0.001, r.accuracy = 0.001, color = "black") +
  coord_cartesian(ylim = c(0,40)) +
  labs(x = "Year", y = "Seed Disperser Abundance \n(no./survey)") +
  theme_bw()+
    theme(
    legend.position = "none",
    axis.text.x = element_text(angle = 30, vjust = 1, hjust=1, size = 14),
    axis.text.y = element_text(size = 14),
    axis.title = element_text(face = "bold", size = 16)
  ))

ggsave("~/GitHub/caplter-birdcommunitydynamics/figures/trends_birds/trend_birds_seeddisp_abundance.png",
       plot.trend,
       width = 6,
       height = 4,
       units = "in",
       dpi = 300)
```


```{r Insectivore trend}
(plot.trend <- data.div %>% 
  ggplot(aes(x = survey_year, y = abund_Invert, group = season, color = season)) +
  geom_point(alpha = 0.3, shape = 16, position = position_dodge(width = 0.6)) +
  geom_smooth(method = "gam") +
  # scale_color_brewer(palette = "Dark2") +
  scale_color_manual(values = pal_season) +
  # stat_cor(method = "spearman", cor.coef.name = "rho", p.accuracy = 0.001, r.accuracy = 0.001, color = "black") +
  coord_cartesian(ylim = c(0,40)) +
  labs(x = "Year", y = "Insectivore Abundance \n(no./survey)") +
  theme_bw()+
    theme(
    legend.position = "none",
    axis.text.x = element_text(angle = 30, vjust = 1, hjust=1, size = 14),
    axis.text.y = element_text(size = 14),
    axis.title = element_text(face = "bold", size = 16)
  ))

# boxplot
(plot.trend <- data.div %>% 
  ggplot(aes(x = as.factor(survey_year), y = abund_Invert, fill = season)) +
    theme_bw() + 
    # geom_point(col = "blue4") +
    geom_boxplot(color = "grey40", outlier.alpha = 0.2, linewidth = 0.3)+
  # scale_fill_brewer(palette = "Dark2") +
  scale_fill_manual(values = pal_season) +
  coord_cartesian(ylim = c(0,40)) +
  labs(x = "Year", y = "Insectivore Abundance \n(no./survey)") +
  theme(
    legend.position = "none",
    axis.text.x = element_text(angle = 40, vjust = 1, hjust=1, size = 14),
    axis.text.y = element_text(size = 14),
    axis.title = element_text(face = "bold", size = 16)
  ))


# lines, points, and error bars
(plot.trend <- data.div %>% reframe(mean = mean(abund_Invert), median = median(abund_Invert),
                                    sd = sd(abund_Invert), se = sd(abund_Invert)/sqrt(length(abund_Invert)), 
                                    length(abund_Invert),
                                    .by = c(survey_year, season)) %>%
  ggplot(aes(x = survey_year, y = mean, group = season, color = season)) +
  geom_errorbar(aes(ymin = (mean-1.96*se), ymax = (mean+1.96*se)), width = 0.4, size = 0.4, 
                position = position_dodge(width = 0.4), alpha = 0.5) +
  geom_point(shape = 16, position = position_dodge(width = 0.4)) +
  geom_line(linewidth = 0.5, position = position_dodge(width = 0.4)) +
  # scale_color_brewer(palette = "Dark2") +
  scale_color_manual(values = pal_season) +
  coord_cartesian(ylim = c(0,40)) +
  labs(x = "Year", y = "Insectivore Abundance \n(no./survey)") +
  theme_bw()+
    theme(
    legend.position = "none",
    axis.text.x = element_text(angle = 30, vjust = 1, hjust=1, size = 14),
    axis.text.y = element_text(size = 14),
    axis.title = element_text(face = "bold", size = 16)
  )
)


ggsave("~/GitHub/caplter-birdcommunitydynamics/figures/trends_birds/trend_birds_invert_abundance.png",
       plot.trend,
       width = 6,
       height = 4,
       units = "in",
       dpi = 300)



```

```{r pollinator trend}
data.div %>% 
  filter(season == "1_winter") %>%
  filter(abund_pollen > 0)

(plot.trend <- data.div %>% 
  ggplot(aes(x = survey_year, y = abund_pollen, group = season, color = season)) +
  geom_point(alpha = 0.3, shape = 16, position = position_dodge(width = 0.6)) +
  geom_smooth(method = "lm") +
  # scale_color_brewer(palette = "Dark2") +
  scale_color_manual(values = pal_season) +
  # stat_cor(method = "spearman", cor.coef.name = "rho", p.accuracy = 0.001, r.accuracy = 0.001, color = "black") +
  coord_cartesian(ylim = c(0, 5)) +
  labs(x = "Year", y = "Pollinator Abundance") +
  theme_bw()+
    theme(
    legend.position = "none",
    axis.text.x = element_text(angle = 30, vjust = 1, hjust=1, size = 14),
    axis.text.y = element_text(size = 14),
    axis.title = element_text(face = "bold", size = 16)
  ))

# boxplot
(plot.trend <- data.div %>% 
  ggplot(aes(x = as.factor(survey_year), y = abund_pollen, fill = season)) +
    theme_bw() + 
    # geom_point(col = "blue4") +
    geom_boxplot(color = "grey40", outlier.alpha = 0.2, linewidth = 0.3)+
  # scale_fill_brewer(palette = "Dark2") +
  scale_fill_manual(values = pal_season) +
  coord_cartesian(ylim = c(0,4)) +
  labs(x = "Year", y = "Pollinator Abundance \n(no./survey)") +
  theme(
    legend.position = "none",
    axis.text.x = element_text(angle = 40, vjust = 1, hjust=1, size = 14),
    axis.text.y = element_text(size = 14),
    axis.title = element_text(face = "bold", size = 16)
  ))


# lines, points, and error bars
(plot.trend <- data.div %>% reframe(mean = mean(abund_pollen), median = median(abund_pollen),
                                    sd = sd(abund_pollen), se = sd(abund_pollen)/sqrt(length(abund_pollen)), 
                                    .by = c(survey_year, season)) %>%
  ggplot(aes(x = survey_year, y = mean, group = season, color = season)) +
  geom_errorbar(aes(ymin = (mean-1.96*se), ymax = (mean+1.96*se)), width = 0.4, size = 0.4, 
                position = position_dodge(width = 0.4), alpha = 0.5) +
  geom_point(shape = 16, position = position_dodge(width = 0.4)) +
  geom_line(linewidth = 0.5, position = position_dodge(width = 0.4)) +
  # scale_color_brewer(palette = "Dark2") +
  scale_color_manual(values = pal_season, labels = c("Winter", "Spring")) +
  # stat_cor(method = "spearman", cor.coef.name = "rho", p.accuracy = 0.001, r.accuracy = 0.001, color = "black") +
  coord_cartesian(ylim = c(0.05,4)) +
  labs(x = "Year", y = "Pollinator Abundance \n(no./survey)", color = "Season") +
  theme_bw()+
    theme(
    # legend.position = "none",
    axis.text.x = element_text(angle = 30, vjust = 1, hjust=1, size = 14),
    axis.text.y = element_text(size = 14),
    axis.title = element_text(face = "bold", size = 16)
  ))


ggsave("~/GitHub/caplter-birdcommunitydynamics/figures/trends_birds/trend_birds_pollinator_abundance.png",
       plot.trend,
       width = 6,
       height = 4,
       units = "in",
       dpi = 300)
```


# Regional species richness trends
```{r}
# adjust based on the % of sites surveyed
div.reg %>% filter(season == "1_winter") %>% pull(rich_reg) %>% range()
div.reg %>% filter(season == "2_spring") %>% pull(rich_reg) %>% range()

max(div.reg$n_site_reg)


(plot.trend <- data.div %>% 
  ggplot(aes(x = survey_year, y = rich_reg, group = season, color = season)) +
  # geom_point(alpha = 0.3, shape = 16, position = position_dodge(width = 0.6)) +
  # geom_smooth(method = "gam") +
  geom_line(linewidth = 0.5) +
  # scale_color_brewer(palette = "Dark2", labels = c("Winter", "Spring")) +
  scale_color_manual(values = pal_season, labels = c("Winter", "Spring")) +
  # scale_fill_brewer(palette = "Dark2") +
  geom_point(alpha = 0.3) +
  # stat_cor(method = "spearman", cor.coef.name = "rho", p.accuracy = 0.001, r.accuracy = 0.001, color = "black") +
  # coord_cartesian(ylim = c(0, 5)) +
  labs(x = "Year", y = "Gamma Diversity \n(no. of species)", color = "Season") +
  theme_bw()+
    theme(
    # legend.position = "none",
    axis.text.x = element_text(angle = 30, vjust = 1, hjust=1, size = 16),
    axis.text.y = element_text(size = 16),
    axis.title = element_text(face = "bold", size = 16)
  ))


ggsave("~/GitHub/caplter-birdcommunitydynamics/figures/trends_birds/trend_birds_regionalrichness.png",
       plot.trend,
       width = 6,
       height = 4,
       units = "in",
       dpi = 300)


(plot.trend <- data.div %>% 
  ggplot(aes(x = survey_year, y = abund_reg_siteavg, group = season, color = season)) +
  # geom_point(alpha = 0.3, shape = 16, position = position_dodge(width = 0.6)) +
  # geom_smooth(method = "gam") +
  geom_line(width = 0.5) +
  geom_point(alpha = 0.3, shape = 16) +
  scale_color_manual(values = pal_season, labels = c("Winter", "Spring")) +
  # stat_cor(method = "spearman", cor.coef.name = "rho", p.accuracy = 0.001, r.accuracy = 0.001, color = "black") +
  # coord_cartesian(ylim = c(0, 5)) +
  labs(x = "Year", y = "Regional Average Total Bird Abundance \n(no. of species)", color = "Season") +
  theme_bw()+
    theme(
    # legend.position = "none",
    axis.text.x = element_text(angle = 30, vjust = 1, hjust=1, size = 16),
    axis.text.y = element_text(size = 16),
    axis.title = element_text(face = "bold", size = 16)
  ))
```







# BONUS
