==> code/3_0_SupplementalMaps_caplter-birdcommunitydynamics.Rmd <==
---
title: "CAP Bird Community Dynamics - Supplementary Visualization of Environmental Conditions"
author: "Jeffrey Haight et al."
date: "`r Sys.Date()`"
output: html_document
---

Code for creating figures visualizing long-term changes in environmental conditions for Appendix S2. Raster data used in visualizations were produced in Google Earth Engine using code from Haight et al. 2024:

Haight, J., De Albuquerque, F., & Frazier, A. (2024). Seasonal and annual summary statistics of urbanization, vegetation, land surface temperature, and bioclimatic variables derived from remotely-sensed imagery in areas surrounding long-term bird monitoring locations in the greater Phoenix, Arizona, USA metropolitan area (1997-2023) [Dataset]. Environmental Data Initiative. https://doi.org/10.6073/PASTA/9D44CD85F881586D6D06E7A7293E833C
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
rm(list = ls())
gc()
```

```{r packages, message = FALSE, echo = TRUE, warning = FALSE}
  library(tidyverse)
  library(beepr)  # for notifying you when some code is  done running

# for spatial data
  library(sf)
  library(terra)
  library(tidyverse)
  library(tidyterra)


# for plotting
# library(ggplot2)      # should already be in the 'tidyverse'
  library(ggpubr)       # for adding correlations and p-values to ggplot
  library(ggeffects)    # for plotting glmm effects in ggplot
  library(ggcorrplot)   # for correlation plots
  library(GGally)       # similar to ggcorrplot
  library(gghighlight)  # for highlighting groups in ggplot
  library(RColorBrewer)
  library(scales)
  library(lterpalettefinder)
  library(basemaps)
```



```{r import study area}

# UTM 12N = ESPG:32612
cap <- st_read("C:/Research/CAPLTER/data/spatial/cap_boundary_Project/cap_boundary_Project.shp") 

# define projections based on these two layers
# crs.aea <- crs(is16.full)
crs.utm <- crs(vect(st_geometry(cap)))

# get extents  
# bb.cap <- st_as_sfc(st_bbox(cap)) %>% st_transform(crs.aea)
# cap.aea <- cap %>% st_transform(crs.aea)

# coerce the CAP site geometry into SpatVector object, for working with SpatRasters
#(cap.mask <- vect(st_geometry(cap)))  # a mask based on just the study area boundary
# (cap.mask <- vect(bb.cap))            # a mask based on the extent of the study area


bird.pts.all <- st_read("C:/Research/CAPLTER/data/core_birds/46_bird_survey_locations_updated.shp") %>% st_transform(crs.utm) %>%
  dplyr::select(-c("begin_date", "begin_mon", "begin_year", "end_date", "end_mon", "end_year")) # drop the time columns, since we'll get this from the bird data itself

bird.pts <- bird.pts.all %>% 
  filter(loc_type %in% c("ESCA", "riparian") )%>%  
  filter(!site_code == "M-9")
```


# Impervious Surface (NDBI) Maps

```{r warning = FALSE}
list.files("C:/Research/CAPLTER/data/spatial/longtermenv/_NDBI")
setwd("C:/Research/CAPLTER/data/spatial/longtermenv/_NDBI")
c(rast("./NDBI_multiseason_CAPLTER_2001.tif"), rast("./NDBI_multiseason_CAPLTER_2002.tif"))[[7]]

ndbi <- c(rast("./NDBI_multiseason_CAPLTER_2001.tif"),#[[-1]],,
         # rast("./NDBI_multiseason_CAPLTER_2002.tif")[[-1]],
         # rast("./NDBI_multiseason_CAPLTER_2003.tif")[[-1]],
         rast("./NDBI_multiseason_CAPLTER_2004.tif"),#[[-1]],
         # rast("./NDBI_multiseason_CAPLTER_2005.tif")[[-1]],
         # rast("./NDBI_multiseason_CAPLTER_2006.tif")[[-1]],
         rast("./NDBI_multiseason_CAPLTER_2007.tif"),#[[-1]],
         # rast("./NDBI_multiseason_CAPLTER_2008.tif")[[-1]],
         # rast("./NDBI_multiseason_CAPLTER_2009.tif")[[-1]],
         rast("./NDBI_multiseason_CAPLTER_2010.tif"),#[[-1]],
         # rast("./NDBI_multiseason_CAPLTER_2011.tif")[[-1]],
         # rast("./NDBI_multiseason_CAPLTER_2012.tif")[[-1]],
         rast("./NDBI_multiseason_CAPLTER_2013.tif"),#[[-1]],
         # rast("./NDBI_multiseason_CAPLTER_2014.tif")[[-1]],
         # rast("./NDBI_multiseason_CAPLTER_2015.tif")[[-1]],
         rast("./NDBI_multiseason_CAPLTER_2016.tif")#[[-1]]
         )
# names(ndbi) <- paste(names(ppt), rep(c("2001","2016"), each = 4), sep = "_")
# names(ndbi) <- paste(c("Winter", "Spring", "Summer", "Fall"), rep(c(2001, 2008, 2016), each = 4), sep = "_")
names(ndbi) <- paste(c("Annual", "Winter", "Spring", "Summer", "Fall"), rep(c(2001, 2004, 2007, 2010, 2013, 2016), each = 5), sep = "_")




```

```{r}
ggplot() +
  # theme_minimal()+
    theme_classic() +
  geom_spatraster(data = ndbi, aes(fill = Winter_2001)) +
  scale_fill_distiller(palette = "Greys", direction = 1, na.value = "transparent", 
                       limits = c(-0.2, 0.4)) +
  geom_sf(data = bird.pts, size = 2.5, col = "white", fill = "transparent", pch = 18) +
  # geom_sf(data = bird.pts, size = 1, col = "white", fill = "transparent") +
  labs(fill = "NDBI") +
  theme(
    strip.background = element_blank(),
    strip.text.x = element_blank(),
    legend.position = "none",
    # axis.title = element_text(face = "bold", size = 16)
    axis.text.x = element_text(angle = 45, vjust = 0.6, size = 8),
    axis.text.y = element_text(size = 8),
    legend.title = element_text(face = "bold")
  ) 
```

```{r}
  
plot.ndbi.all <- ggplot() +
  # theme_minimal()+
    theme_classic() +
  geom_spatraster(data = ndvi) +
  facet_wrap(~lyr, ncol = 4) +
  scale_fill_distiller(palette = "Greys", direction = 1, na.value = "transparent", 
                       limits = c(-0.2, 0.4)) +
  # geom_sf(data = bird.pts, size = 0.3, col = "black", fill = "transparent") +
  # geom_sf(data = bird.pts, size = 0.05, col = "white", fill = "transparent") +
  labs(fill = "NDBI") +
  theme(
    strip.background = element_blank(),
    strip.text.x = element_blank(),
    legend.position = "none",
    # axis.title = element_text(face = "bold", size = 16)
    axis.text.x = element_text(angle = 45, vjust = 0.6, size = 8),
    axis.text.y = element_text(size = 8),
    legend.title = element_text(face = "bold")
  ) 

# plot.ndvi.all
ggsave("~/GitHub/caplter-birdcommunitydynamics/figures/trends_env/map_NDBI_multiyear.png",
       plot.ndbi.all,
       width = 6,
       height = 6,
       units = "in",
       dpi = 300)

beep()
```

# Vegetation (NDVI) Maps
```{r}
list.files("./NDVI")
c(rast("./NDVI/NDVI_multiseason_CAPLTER_2001.tif"), rast("./NDVI/NDVI_multiseason_CAPLTER_2002.tif"))[[7]]

ndvi <- c(rast("./NDVI/NDVI_multiseason_CAPLTER_2001.tif")[[-1]],
         # rast("./NDVI/NDVI_multiseason_CAPLTER_2002.tif")[[-1]],
         # rast("./NDVI/NDVI_multiseason_CAPLTER_2003.tif")[[-1]],
         rast("./NDVI/NDVI_multiseason_CAPLTER_2004.tif")[[-1]],
         # rast("./NDVI/NDVI_multiseason_CAPLTER_2005.tif")[[-1]],
         # rast("./NDVI/NDVI_multiseason_CAPLTER_2006.tif")[[-1]],
         rast("./NDVI/NDVI_multiseason_CAPLTER_2007.tif")[[-1]],
         # rast("./NDVI/NDVI_multiseason_CAPLTER_2008.tif")[[-1]],
         # rast("./NDVI/NDVI_multiseason_CAPLTER_2009.tif")[[-1]],
         rast("./NDVI/NDVI_multiseason_CAPLTER_2010.tif")[[-1]],
         # rast("./NDVI/NDVI_multiseason_CAPLTER_2011.tif")[[-1]],
         # rast("./NDVI/NDVI_multiseason_CAPLTER_2012.tif")[[-1]],
         rast("./NDVI/NDVI_multiseason_CAPLTER_2013.tif")[[-1]],
         # rast("./NDVI/NDVI_multiseason_CAPLTER_2014.tif")[[-1]],
         # rast("./NDVI/NDVI_multiseason_CAPLTER_2015.tif")[[-1]],
         rast("./NDVI/NDVI_multiseason_CAPLTER_2016.tif")[[-1]]
         )
# names(ppt) <- paste(names(ppt), rep(c("2001","2016"), each = 4), sep = "_")
# names(ndvi) <- paste(c("Winter", "Spring", "Summer", "Fall"), rep(c(2001, 2008, 2016), each = 4), sep = "_")
names(ndvi) <- paste(c("Winter", "Spring", "Summer", "Fall"), rep(c(2001, 2004, 2007, 2010, 2013, 2016), each = 4), sep = "_")



ggplot() +
  # theme_minimal()+
    theme_classic() +
  geom_spatraster(data = ndvi, aes(fill = Winter_2001)) +
  scale_fill_viridis_c(option = "viridis", direction = 1, na.value = "transparent", limits = c(-0.0, 0.7), oob = scales::squish) +
  geom_sf(data = bird.pts, size = 0.8, col = "black", fill = "transparent") +
  geom_sf(data = bird.pts, size = 0.5, col = "white", fill = "transparent") +
  labs(fill = "NDVI") +
  theme(
    strip.background = element_blank(),
    strip.text.x = element_blank(),,
    # axis.title = element_text(face = "bold", size = 16)
    axis.text.x = element_text(angle = 45, vjust = 0.6, size = 8),
    axis.text.y = element_text(size = 8),
    legend.title = element_text(face = "bold")
  ) 
```

```{r}
  
plot.ndvi.all <- ggplot() +
  # theme_minimal()+
    theme_classic() +
  geom_spatraster(data = ndvi) +
  facet_wrap(~lyr, ncol = 4) +
  scale_fill_viridis_c(option = "viridis", direction = 1, na.value = "transparent", limits = c(-0.0, 0.7), oob = scales::squish) +
  # geom_sf(data = bird.pts, size = 0.3, col = "black", fill = "transparent") +
  # geom_sf(data = bird.pts, size = 0.05, col = "white", fill = "transparent") +
  labs(fill = "NDVI") +
  theme(
    strip.background = element_blank(),
    strip.text.x = element_blank(),
    legend.position = "none",
    # axis.title = element_text(face = "bold", size = 16)
    axis.text.x = element_text(angle = 45, vjust = 0.6, size = 8),
    axis.text.y = element_text(size = 8),
    legend.title = element_text(face = "bold")
  ) 

# plot.ndvi.all
ggsave("~/GitHub/caplter-birdcommunitydynamics/figures/trends_env/map_NDVI_multiyear.png",
       plot.ndvi.all,
       width = 6,
       height = 6,
       units = "in",
       dpi = 300)
```

# Land Surface Temperature (LST) Maps
```{r}
list.files("./LST")
c(rast("./LST/LST_multiseason_CAPLTER_2001.tif"), rast("./LST/LST_multiseason_CAPLTER_2002.tif"))[[7]]

LST <- c(rast("./LST/LST_multiseason_CAPLTER_2001.tif")[[-1]],
         # rast("./LST/LST_multiseason_CAPLTER_2002.tif")[[-1]],
         # rast("./LST/LST_multiseason_CAPLTER_2003.tif")[[-1]],
         rast("./LST/LST_multiseason_CAPLTER_2004.tif")[[-1]],
         # rast("./LST/LST_multiseason_CAPLTER_2005.tif")[[-1]],
         # rast("./LST/LST_multiseason_CAPLTER_2006.tif")[[-1]],
         rast("./LST/LST_multiseason_CAPLTER_2007.tif")[[-1]],
         # rast("./LST/LST_multiseason_CAPLTER_2008.tif")[[-1]],
         # rast("./LST/LST_multiseason_CAPLTER_2009.tif")[[-1]],
         rast("./LST/LST_multiseason_CAPLTER_2010.tif")[[-1]],
         # rast("./LST/LST_multiseason_CAPLTER_2011.tif")[[-1]],
         # rast("./LST/LST_multiseason_CAPLTER_2012.tif")[[-1]],
         rast("./LST/LST_multiseason_CAPLTER_2013.tif")[[-1]],
         # rast("./LST/LST_multiseason_CAPLTER_2014.tif")[[-1]],
         # rast("./LST/LST_multiseason_CAPLTER_2015.tif")[[-1]],
         rast("./LST/LST_multiseason_CAPLTER_2016.tif")[[-1]]
         )
# names(LST) <- paste(names(ppt), rep(c("2001","2016"), each = 4), sep = "_")
# names(LST) <- paste(c("Winter", "Spring", "Summer", "Fall"), rep(c(2001, 2008, 2016), each = 4), sep = "_")
names(LST) <- paste(c("Winter", "Spring", "Summer", "Fall"), rep(c(2001, 2004, 2007, 2010, 2013, 2016), each = 4), sep = "_")



ggplot() +
  # theme_minimal()+
    theme_classic() +
  geom_spatraster(data = LST, aes(fill = Winter_2001)) +
  scale_fill_distiller(palette = "RdYlBu", direction = -1, na.value = "transparent", 
                       limits = c(10, 60), oob = scales::squish) +
  # geom_sf(data = bird.pts, size = 0.8, col = "black", fill = "transparent") +
  # geom_sf(data = bird.pts, size = 0.5, col = "white", fill = "transparent") +
  labs(fill = "LST (°C)") +
  theme(
    strip.background = element_blank(),
    strip.text.x = element_blank(),,
    # axis.title = element_text(face = "bold", size = 16)
    axis.text.x = element_text(angle = 45, vjust = 0.6, size = 8),
    axis.text.y = element_text(size = 8),
    legend.title = element_text(face = "bold")
  ) 
```


```{r}
  
plot.LST.all <- ggplot() +
  # theme_minimal()+
    theme_classic() +
  geom_spatraster(data = LST) +
  facet_wrap(~lyr, ncol = 4) +
  scale_fill_distiller(palette = "RdYlBu", direction = -1, na.value = "transparent", 
                       limits = c(10, 60), oob = scales::squish) +
  # geom_sf(data = bird.pts, size = 0.3, col = "black", fill = "transparent") +
  # geom_sf(data = bird.pts, size = 0.05, col = "white", fill = "transparent") +
  labs(fill = "LST (°C)") +
  theme(
    strip.background = element_blank(),
    strip.text.x = element_blank(),
    legend.position = "none",
    # axis.title = element_text(face = "bold", size = 16)
    axis.text.x = element_text(angle = 45, vjust = 0.6, size = 8),
    axis.text.y = element_text(size = 8),
    legend.title = element_text(face = "bold")
  ) 

# plot.ndvi.all
ggsave("~/GitHub/caplter-birdcommunitydynamics/figures/trends_env/map_LST_multiyear.png",
       plot.LST.all,
       width = 6,
       height = 6,
       units = "in",
       dpi = 300)
```

# Precipitation Maps
```{r import precipitation data}

list.files("./Daymet/ppt_sum")
ppt <- c(rast("./Daymet/ppt_sum/ppt_sum_multiseason_CAPLTER_Daymet_2001.tif")[[-1]],
         # rast("./Daymet/ppt_sum/ppt_sum_multiseason_CAPLTER_Daymet_2002.tif")[[-1]],
         # rast("./Daymet/ppt_sum/ppt_sum_multiseason_CAPLTER_Daymet_2003.tif")[[-1]],
         rast("./Daymet/ppt_sum/ppt_sum_multiseason_CAPLTER_Daymet_2004.tif")[[-1]],
         # rast("./Daymet/ppt_sum/ppt_sum_multiseason_CAPLTER_Daymet_2005.tif")[[-1]],
         # rast("./Daymet/ppt_sum/ppt_sum_multiseason_CAPLTER_Daymet_2006.tif")[[-1]],
         rast("./Daymet/ppt_sum/ppt_sum_multiseason_CAPLTER_Daymet_2007.tif")[[-1]],
         # rast("./Daymet/ppt_sum/ppt_sum_multiseason_CAPLTER_Daymet_2008.tif")[[-1]],
         # rast("./Daymet/ppt_sum/ppt_sum_multiseason_CAPLTER_Daymet_2009.tif")[[-1]],
         rast("./Daymet/ppt_sum/ppt_sum_multiseason_CAPLTER_Daymet_2010.tif")[[-1]],
         # rast("./Daymet/ppt_sum/ppt_sum_multiseason_CAPLTER_Daymet_2011.tif")[[-1]],
         # rast("./Daymet/ppt_sum/ppt_sum_multiseason_CAPLTER_Daymet_2012.tif")[[-1]],
         rast("./Daymet/ppt_sum/ppt_sum_multiseason_CAPLTER_Daymet_2013.tif")[[-1]],
         # rast("./Daymet/ppt_sum/ppt_sum_multiseason_CAPLTER_Daymet_2014.tif")[[-1]],
         # rast("./Daymet/ppt_sum/ppt_sum_multiseason_CAPLTER_Daymet_2015.tif")[[-1]],
         rast("./Daymet/ppt_sum/ppt_sum_multiseason_CAPLTER_Daymet_2016.tif")[[-1]]
         )
# names(ppt) <- paste(names(ppt), rep(c("2001","2016"), each = 4), sep = "_")
# names(ppt) <- paste(c("Winter", "Spring", "Summer", "Fall"), rep(c(2001, 2008, 2016), each = 4), sep = "_")
names(ppt) <- paste(c("Winter", "Spring", "Summer", "Fall"), rep(c(2001, 2004, 2007, 2010, 2013, 2016), each = 4), sep = "_")
 
ggplot() +
  # theme_minimal()+
    theme_classic() +
  geom_spatraster(data = ppt, aes(fill = Winter_2001)) +
  scale_fill_viridis_c(option = "viridis", direction = -1, 
                       na.value = "transparent", limits = c(0, 250), oob = scales::squish) +
  # geom_sf(data = cap, col = "black", fill = "transparent") +
  geom_sf(data = bird.pts, size = 0.5, col = "black", fill = "transparent") +
  labs(fill = "Precipitation \n(mm)") +
  theme(
    strip.background = element_blank(),
    strip.text.x = element_blank(),
    axis.text.x = element_text(angle = 45, vjust = 0.6)
  )
  
 
  
```
```{r}
plot.ppt.all <- ggplot() +
  # theme_minimal()+
    theme_classic() +
  geom_spatraster(data = ppt) +
  facet_wrap(~lyr, ncol = 4) +
  scale_fill_viridis_c(option = "viridis", direction = -1, 
                       na.value = "transparent", limits = c(0, 250), oob = scales::squish) +
  # geom_sf(data = bird.pts, size = 0.3, col = "black", fill = "transparent") +
  # geom_sf(data = bird.pts, size = 0.05, col = "white", fill = "transparent") +
  labs(fill = "Precipitation \n(mm)") +
  theme(
    strip.background = element_blank(),
    strip.text.x = element_blank(),
    legend.position = "none",
    # axis.title = element_text(face = "bold", size = 16)
    axis.text.x = element_text(angle = 45, vjust = 0.6, size = 8),
    axis.text.y = element_text(size = 8),
    legend.title = element_text(face = "bold")
  ) 
# , limits = c(0, 400)
# 
# ggplot() +
#   # theme_minimal()+
#     theme_classic() +
#   geom_spatraster(data = ppt) +
#   facet_wrap(~lyr, ncol = 4) +
#   scale_fill_viridis_c(option = "viridis", direction = -1, na.value = "transparent") +
#   # geom_sf(data = cap, col = "black", fill = "transparent") +
#   geom_sf(data = bird.pts, size = 0.5, col = "black", fill = "transparent") +
#   labs(fill = "Precipitation \n(mm)") +
#   theme(
#     strip.background = element_blank(),
#     strip.text.x = element_blank(),
#     axis.text.x = element_text(angle = 40, vjust = 0.6)
#   )

ggsave("~/GitHub/caplter-birdcommunitydynamics/figures/trends_env/map_ppt_multiyear.png",
       plot.ppt.all,
       width = 6,
       height = 6,
       units = "in",
       dpi = 300)
```
